addressed some of my comments
