+++ pennylane/labs/resource_estimation/ops/op_math/symbolic.py
@@ -728,6 +728,9 @@ class ResourcePow(PowOperation, re.ResourceOperator):
         if z == 0:
             return {}
 
+        if z == 1:
+            return {base_class.resource_rep(**base_params): 1}
+
         try:
             return base_class.pow_resource_decomp(z, **base_params)
         except re.ResourcesNotDefined:
